ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
keras-ocr==0.8.9
sympy==1.12
numpy>=1.21.0,<2.0.0
Pillow>=9.0.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0